+++ model/src/main/scala/com/codefactory/ecommerce/tableModel/CartProduct.scala
@@ -1,0 +1,3 @@
+package com.codefactory.ecommerce.tableModel
+
+case class CartProduct(id: Int, c_id: Int, p_id: Int)