+++ persistence/src/main/scala/com/codefactory/ecommerce/tableVariables/TableVariables.scala
@@ -30,7 +30,10 @@ trait TableVariables {
   final val cartProductStatus = "CART_PRODUCT_STATUS"
   final val cartTotal = "CART_TOTAL"
   final val userIDcart = "CART_USER_ID"
-  final val cartProductId = "CART_PRODUCT_ID"
   //END
 
+  // CARTPRODUCT TABLE VARIABLES - RELATIONSHIP BETWEEN THE TWO
+  final val cartProductID = "CART_PRODUCT_ID"
+  final val cartId = "C_ID"
+  final val productId = "P_ID"
 }