+++ src/main/resources/db/migration/V2__INSERT INTO TABLES.sql
@@ -6,3 +6,12 @@ INSERT INTO PRODUCTS VALUES
 (1,'Cana',10,'vesela',5,1),
 (2,'Vaza',30,'decoratiune',16,0),
 (3,'Pian',1000,'muzica',3,1);
+
+INSERT INTO CART VALUES
+(1,3,'In Stock',1,300),
+(2,2,'Out of stock',1,200);
+
+INSERT INTO CART_PRODUCT VALUES
+(1,1,1),
+(2,1,2),
+(3,1,3);